backend/requirements/local.txt
# Import all base requirements
-r base.txt

# Django debug toolbar
django-debug-toolbar==1.9.1

# Code linter
flake8==3.6.0
